refactor: reduce authority in Rholang grammar handlng
`function setRholangHighlight(Prism)` had the power to completely
redefined everything about Prism. The rholang grammar definition is
just a big JSON constant. Defining it needs no authority at all.
So let's just define it and export it as a constant.

index.js is our exception from ocap discipline. As such, it should do
as little as possible. It should only get access to the outside world
and pass that access to ocap-happy code for non-trivial processing. So
rather than calling `setRholangHighlight(Prism)` in index.js, let's
just make a `setGrammar` function that can set the highlighting
grammar for a language and pass that to buildUI. Note that `index.js`
need not know the details of the grammar nor the name of the
language. (This is perhaps not the least authority we could provide to
buildUI. We could pass a `setRholangGrammar` function with the
language name hard-coded to 'rholang'.)

Now that RholangGrammar is just a constant, we can import it from
participate.js and then call the provided `setLanguage` from `buildUI`.