refactor: give buildUI access to syntax highlighting operations
buildUI() was accessing the powerful `document` ambiently. We could
pass that power in explicitly, but power to update the whole document
is a lot more than buildUI needs. So we hoist the updateHighlighting
and syncScroll functions to index.js and grant just those operations
to buildUI.

Calls to updateHighlighting were paired with state.term = ... in
various places. Instead, let's put the call to updateHighlighting in
the setter for state. We schedule it asynchronously to let the DOM
updates happen first. (lightly tested)